emitter: Enums now entered as functions without writing code.
This allows enums to be consistent with classes. It should make some
internal optimizations possible in the near future. Also, note that
enums are no longer setting the emitter's class entry. That's fine
because the class entry is only used to check private/protected method
access, and enum methods are always public.