tooling: Create new tooling to generate lily_token and lexer tables.
Prior to this, lexer was allocating and initializing a table that
lily_next_token used for grouping. That table is now generated for the
lexer.

The lexer's identity table is now generated and the token to string
table in tokname is also generated. The latter was done to ensure that
token descriptions don't misalign by accident (it's happened before).

The lily_token enum is now generated. It's now in a file by itself in
case other parts of the interpreter want to use it in the future.