parser,tooling: Generated dispatch table for expressions (fixes #458).
Most tokens in expression are complicated enough that they need a
function to handle them. Now all of them are handled in a function
that's put into a dispatch table. The dispatch table is generated so
that it's easier to verify (and also easier to change if that becomes
necessary).

This made `src/lily_parser_tok_table.h` and the lily_expr_op enum
obsolete, so both have been removed.