parser: Block handling is no longer recursive.
The line count is small because this patch builds on recent work, but
this is a massive rethink of block parsing.

Blocks are keywords that are entered, like `import`, `if`, `while` and
so on. Recursive block handling means that there's a function that's
in scope as long as the block is in scope.

With this patch, keyword functions now do block setup and return.
Lambda parsing and parser loop are now responsible for exiting out of
blocks and making sure curly braces are balanced.

Some tests were changed since this resulted in line numbers being off.
Since keywords were being moved around anyway, error messages were
changed to no longer quote keywords in hopefully all cases. It was
being done inconsistently, and didn't seem all that useful.