manifest: It's finally over. The prelude now uses manifest files.
This commit introduces the `predefined` keyword, allowing the builtin
manifest to override builtin classes and enums. Along with that, it
also adds an override for `builtin` for the library keyword. Those two
changes were the last barriers to migration.

All modules in the prelude have been rewritten to use manifest files
for generating their bindings.