emitter: Rewrite the start of calls to fix a bug (fixes #468).
The old call start was drawing the call type from the item. That's the
correct move to make except with properties. For a property, the item is
the unsolved property, and the right type is the result's type.

Rewriting calls also tames some of the mess that starting a call became
over time.