Compare and equality opcodes now include a jump (fixes #406).
In many cases, comparison and equality checks are the base of an `if`,
`elif`, or `while` entry check. Those operations now include a jump that
they take if the comparison fails (jump if false). This includes basic
folding if a conditional op is the root tree to eliminate a jump.

If a comparison or equality tree is not the root, a storage is loaded
with 1 if true, 0 if false. This is a slowdown, but not one which should
be run into frequently.

This was inspired by seeing Lua do the same. Testing on fibonacci shows
a consistent and noticeable speed boost. The testing suite seems to be
a tiny bit faster too.