Massive rewrite of many prelude methods (fixes #485).
This rewrite is a long time coming. The rewrite focuses primarily on
`List` and `String` methods, which were written a long time ago. Methods
of other classes have been given a few style adjustments, but are
otherwise the same.

This was done with the aim of clearing out msvc warnings. The highlights
are as follows:

* Methods for `List` and `String` now use api functions to grab data and
  size.

* Several `String` methods have been simplified by using pointers and C
  standard library functions.

* Methods for `List` should now properly wrap around when given a
  negative `Integer`.

* The test suite now has a few more assertions for less common `List`
  and `String` method situations.