Add basic support for constants (fixes #436).
This allows the declaration of basic constants of `Double`, `Integer`,
and `String`. Declaration is limited to a single token to make constants
easy to implement.

Support for constants is currently limited to native code. Introspection
provides functions for extracting information about constants. However,
manifest mode (and thus dynaload) do not observe constants yet.