Modest expansion of test coverage into uncommon areas.
This patch adds extra coverage to less-used parts of the interpreter,
like the string pile, the uncommon parts of the type system, and so on.

On my machine, coverage with this patch is at 95%. Given the complexity
of an interpreter, 100% coverage is not a sign of a lack of defects.
Several functions are flagged as having an unreachable line of code due
to calling a raise function, so 100% is not possible anyway.

Raising coverage in the interpreter makes it easier to identify what
code is not necessary, and helps prevents accidental regressions.