Do not interpolate `Byte` or `ByteString` when printing (fixes #561).
This changes the output of `print`, `File.write`, and
`File.write_to_path`. This is technically a breaking change, but I am
okay with that because the old behavior was absolute nonsense.

Prior to this patch, attempting to print a `ByteString` or `Byte` value
resulted in the value being interpolated first. Such behavior was an
oversight, and not at all a reasonable expectation: Attempting to write
`'x'` would result in a file saying, literally `'x'` and not just `x`.
`ByteString` values had their `B"..."` encasing around their content.

This patch does not alter the representation of `ByteString` or `Byte`
when inside of a container (such as a `Some` or a `List`). Such behavior
matches how `String` is treated in such situations.