Stricter limits for globals with generics or scoop (fixes #554).
Consider this example:

```
define identity[A](a: A): A { return a }

define f[A](g: Function(A => A)) {}
```

The interpreter's type system sees `identity` and `g` as being
equivalent to each other. I don't think the benefits of teaching the
interpreter the difference are worth the trouble.

Global definitions that use scoop (`$1`) such as `List.zip` must now be
immediately called. This prevents `var xyz = [List.zip]`, but that is
nowhere near reasonable.

If a global definition uses generics, it must either be called or a call
argument (where it will be shape matched). This is to allow
`xyz.map(Option.unwrap)` which I *do* consider to be reasonable, but
prohibits `[1 => Option.unwrap]`, which is entirely unreasonable.

Globals no longer need to be explicitly prevented from being assigned a
global generic, because the above should make that impossible.