Internally represent generics as `lily_generic_class`.
Prior to this patch, the generic pool stored generics as a full class
and created a unique type for each.

This patch introduces `lily_generic_class`. It's a slightly slimmed
version of `lily_class`. It's a type for itself, so no more creating a
new `lily_type` as well. The overall effect is a small reduction in
memory usage.

Additionally, some shuffling has been done with generics. It's now
possible to do `type->cls_id` instead of `type->cls->id`. Scripts should
execute exactly as they did before.