Parser no longer calls `lily_next_token` before dispatching a keyword.
If most keywords will want to pull the next token, why not call for it
before dispatching? That line was the prior thinking. In most cases,
that thinking was correct.

Suppose `public` is found outside of a class. The next token has already
been pulled up. That might have changed the line number. It's no longer
possible (at least, not easily so) to point to the correct identifier as
the problem.

Now, various keywords are responsible for calling for their first token.
A few tests have been updated as their line numbers are now correct.