Some lexer errors now include context (#559).
This commit introduces a framework for showing the current line when an
error occurs.

In areas where context does not make sense (unterminated tokens), the
lexer sets a new `token_start` to NULL to indicate that no trace should
be shown.

As of this commit, only lexer uses this framework. Future commits will
extend this new framework to parser and emitter.