Allow forward classes in manifest mode (fixes #568).
This introduces forward classes to the interpreter's manifest mode. They
are available using `forward class <name>[generics] { ... }`.

Classes must be resolved with the same amount of generics. However, it
is not necessary to use the same manifest qualifiers: You can have a
`forward class Example { ... }` resolved as a `foreign class`, a
`foreign static class` or just a `class`.