Most parser errors now include context (#559).
This commit gives parser access to the framework for showing the current
line when an error occurs. Parser has areas where it reads the token too
far, so the cursor isn't at the best spot. However, it's still in the
area, and the line is still shown, which are the important parts.

Lexer now saves and restores where the token starts. There are some edge
cases with dynaload and some with lambdas which make that necessary.

Context is now shown if it would point to a blank line with a closing
curly brace. That's done to prevent showing useless context information.
That can happen if a match is not complete, or a definition doesn't
return a value at the end.