Expressions now allow variant shorthand when using an enum (#575).
This does not fix the above issue, because it does not alter how `match`
works. That will be addressed shortly in a different commit.

If an expression expects an enum and `.<somevariant>` is given, the enum
portion is inferred. This is primarily intended to make api that uses
enums from another module more palatable. With that goal in mind, the
syntax is intentionally the same and enabled for both scoped and flat
enums.