Add UTF-8 library
This adds a basic UTF-8 library for working with codepoints,
functionality that was previously missing from Strings. Only simple
codepoint retrieval and iteration is implemented - grapheme clusters are
too complicated to be worth handling, and any kind of character
inspection would require large tables of data from the Unicode Character
Database (though, the size of said tables depends on what exact
functionality is desired). Tests and documentation are included.