Introduce value enums and value variants (fixes #581).
A value enum is an enum that inherits from `Integer`. The variants of a
typical enum are unique entities (why `print(None)` prints None). Value
variants, on the other hand, operate akin to C enums: Each one has a
value (must be an `Integer` literal) and no type information.

These have been implemented by having the enum actually inherit from
`Integer`. Since this is the only time a primitive has been inherited
from, several tests have been added to ensure no silly behavior occurs.

If a value variant is not provided a value, it will default to the prior
variant's value `+1`. If no value is given for the first variant, the
default is `0`. Value variants do not permit duplicates. It is assumed
that duplicates are likely an error. Also, it would make `match`
difficult.

Value enums are provided all the features of typical enums: They can be
scoped or not, allow methods, `match`, and `with`, typecast, and can be
in manifest files.