Add methods to generate random Doubles
This adds two new methods to the random.Random class:

- Random.double: Generates a random Double in the range [0, 1).
- Random.double_between: Generates a random Double between two other
  values.

It wasn't impossible to generate random Doubles before, but it required
extra work to do so with only an Integer generator. This makes it much
more convenient.

Documentation and tests are included. I also added a new test for
determinism between random.Random objects given the same seed, as I felt
that was missing before (if you'd rather that not be included here, it
can be easily removed).

Hopefully I didn't miss anything like last time.