Allow direct import of modules (fixes #582).
When direct import was added (675f300f, from 2017), all symbols except
for modules were allowed in. No explanation was given for why.

When module A wants to direct import C from B, it must use the name that
B uses for C. If B says `import C as abc`, then A must
`import (abc) B`. This is consistent with module renaming. This commit
does not allow A to rename `abc` again.

For introspection: Directly imported modules are available in the list
of modules that are used. There is no `boxed_modules`.