Allow `for i(, element) in list` (fixes #571).
The `for` keyword can now take a `List` to iterate over, in addition to
the range it allowed before.

When `for` is given a `List`, it can produce either elements, or indexes
and elements. The new style can use existing variables just as does now.

The new style has two different behaviors than a for over a range.
First, it does not allow global variables to be used as iterators.
Second, modifying the iterator will change the loop position. A future
patch will add those behaviors to the existing style of for.

This patch makes `foreach` obsolete. To avoid breaking existing code, it
has been left alone for the time being.