Allow ranges of length 1 in Random.double_between
This updates Random.double_between to match the new behavior of
Random.between, where passing in a range of length 1 (ie. the low and
high values are the same) will always return the single number contained within that range.

Additionally, an explicit check for this case has been added to both
methods, to ensure this behavior is consistent/reliable. It's probably
not strictly necessary, particularly for integers, but for doubles I was
concerned about floating point error potentially causing problems, and
so I felt it was best form to add it to both methods.