Introduce Byte -> Integer autopromotion (#584).
The interpreter has, until now, always treated numeric types as being
distinct from each other. This commit introduces implicit conversion
of Byte to Integer, termed autopromotion.

If an eval function produces a Byte where an Integer is expected, the
result is autopromoted. Functions that don't want autopromotion can
supply `?` as inference.

Autopromotion in vm requires a couple factors: The value must be cloned,
and the value must be marked as an Integer. Autopromotion is implemented
by taking a value (ex: `x`) and performing `(storage) = x | x`. This
approach was chosen because it is assumed to be faster than creating a
new opcode to copy + mark the value.