More ergonomic math operations (#584).
Prior to this commit, Lily's math operations had a strict rule: Both
sides must be Integer or Double. That rule was never revised, even with
the introduction of Byte.

Math operations are now allowed between any mix of Byte, Integer, and
Double (except Double and Byte). The result of a math operation is a
Double if either side is a Double, or Integer. By converting Byte to
Integer in operations, they can't easily overflow.

If either side is a Double (the other being Integer), the value is
promoted before the math occurs. This new Integer -> Double promotion is
limited to math primitives for the time being.