Fix possible stale references to quicksort pivots
Previously, the pivot value in the quicksort function was just stored as
a lily_value *, which - from what I suspect, anyway - meant the value it
pointed to could change after a swap. That could cause sorted results to
be incorrect in some cases. I was able to fix this by pushing the value
to the stack, and retrieving it before each iteration. I'm not sure if
that's necessarily the most efficient way to fix this, but it was the
easiest solution I could think of.