The last match branch of an enum always succeeds (fixes #593).
The last branch of a match on an enum will always succeed, so don't
check the value. The check was trying to jump to after the last match
block. If all branches of the match block returned, the jump goes to an
invalid (but unreachable) section. Closure transformation was doing an
invalid read, since it follows jumps.