Fix misbehavior when a Coroutine takes a closure (#595).
Coroutines make a private copy of a Function when they start. If that
Function happened to be a closure, it was copying the gc lock. I was
able to put together a test that shows a leak in valgrind with the old
behavior.

I suspect this fixes the root cause of the above issue. If not, I think
the two are at least related.