Make executable and tests optional in CMake
By default, Lily builds a standalone executable and test suite in
addition to liblily itself. This is desired in most cases, but when eg.
building Lily as a subdirectory inside another CMake project, it can be
wasteful. This makes both the executable and test suite able to be
disabled through CMake options (LILY_BUILD_EXECUTABLE and
LILY_BUILD_TESTS). Both options are enabled by default, so existing
builds shouldn't change.

Additionally, this commit fixes the "check" CMake target, which was
previously broken. It appeared to have been written for an older
version of Lily without a standalone pre-commit-tests executable, as it
needlessly attempted to copy the standard Lily executable into the
top-level directory, which failed on at least my system. The target has
been fixed to just run pre-commit-tests (assuming that's all it was
meant to do originally). I noticed the issue because the part of the
script I was already modifying also involved the target in question.