Allow List, String, and ByteString subscripts by value variants.
This was made possible by the machinery added in the last commit. This
leaves out Hash and Tuple. Making this work for Hash is somewhat
involved, so I'll leave it for another time. I don't intend to add this
to Tuple unless asked, since I doubt it will be useful.