Class constructors cannot be a closure origin (fixes #599).
The included test was crashing because the class method thought it was
the backing closure, and scooped the upvalues before the class could
scoop them. Classes are blocked from closures because that has been the
source of many problems in the past (like this one!).

By marking the lambda as the backing closure, it scoops the upvalues
before the method gets them.