Add BUILD_SHARED_LIBS CMake option
Previously, liblily was always built as a static library on Windows, and
as a shared library on all other platforms. This was sufficient most of
the time, but it is nice to be able to choose what kind of library you
want to build. The majority of libraries using CMake I've seen have such
an option, but Lily was lacking one.

Different libraries handle this in different ways, but the simplest
solution seemed to be to use the BUILD_SHARED_LIBS option, which is
decently common and officially supported by CMake (if the option is
present, CMake will use it to automatically select shared/static for
libraries that don't explicitly specify one or the other). This commit
updates Lily's CMake scripts to take this option into account. The
default is to build as a shared library, which seems to be what most
libraries do.

The reason liblily was built as a static library on Windows originally
was that, if it wasn't, C extensions would not be able to find the
symbols they needed if they used any interpreter functions. Indeed, when
I tried building as a shared library on Windows as-is, an import library
wasn't even created, and so C extensions would fail to find Lily when
attempting to build. I was able to resolve this by setting the
CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS variable to ON when building as a
shared library on Windows.

Additionally, I replaced the use of the -fPIC compiler flag with the
equivalent (but more idiomatic, I think?) POSITION_INDEPENDENT_CODE
CMake property. There shouldn't be any functional difference there,
though.