Remove template mode (fixes #579).
Lily's original design was as a template language. Over time, focus has
shifted toward being a general purpose language. The web has also
changed tremendously since Lily was started. This functionality is not
a large part of the interpreter, but it is not being used and I do not
suspect I will regain interest in it. Thus, I am removing it.