Line context for emitter errors on the current line (#604).
Expression trees now carry a token start that error building can use.
This expands error context reporting to error building, but only if the
error is on the current line. This may seem useful, but in most cases it
will not be. Since parser pulls the next token to stop an expression,
this is unlikely to be hit. This, therefore, is more of a
proof-of-concept for the real context work.