Add Coroutine.status method (fixes #602).
A Coroutine can exist in one of four possible states. Prior to this, the
only mechanism for checking the status of a Coroutine was to call four
different `is_*` methods to see which returned true.

The new method returns a CoStatus scoped enum that has one variant for
each possible Coroutine state.