Fix Random's constructor ignoring seed arguments
The random.Random constructor erroneously checked the second argument
slot for a seed, instead of the first (perhaps assuming the first slot
would be taken up by a Random object, but that's not true for
constructors!). This meant that passing a seed to the constructor didn't
actually do anything. I've replaced the broken code with a call to
lily_optional_integer instead, which makes the constructor behave as
documented.