Add UTF-8 encoding functions
The `utf8` module works with codepoints, not `String`s. That is the most
efficient way to handle things, as using `String`s would require a lot
of extraneous allocations, but it can be inconvenient for programs
wanting to work with actual text instead of numbers. Up until now, the
module allowed converting from `String`s to codepoints, but not the
other way around, forcing programs to do the encoding themselves.

To resolve this, I've added two new functions to the `utf8` module:
- `utf8.encode`: Encodes an `Integer` codepoint as UTF-8. Returns a
  `Some(String)` on success, and `None` on failure (ie. if the codepoint
  is invalid).
- `utf8.encode_list`: Encodes each codepoint in a `List` as UTF-8, and
  joins them together into a single `String`. The return values work the
  same way as in `utf8.encode`. This function serves as a direct inverse
  to the existing `utf8.as_list`.

Documentation and tests are included. The encoding logic is in a static
function in `lily_pkg_utf8.c` right now, instead of in `lily_utf8.c`, as
it's currently not needed anywhere else. If that changes for whatever
reason, it could easily be moved.