Coroutine resumption now returns a Result (fixes #612).
This adds a new exception, CoError, to the coroutine module. If
resumption fails because of a usage error, a CoError is returned in the
Failure with the caller's traceback. In all other cases, the exception
(if any) has the trace of the Coroutine being used.

This patches the interpreter's dynaload mechanism to load CoError when
Coroutine is loaded. That is not ideal, but it avoids extending vm
exception dynaload, which I strongly prefer to keep as confined and
minimal as possible.

This **does not** add an API function to raise CoError, because that
exception should only be used by Coroutine resumption.