Add `utf8.each_codepoint_with_index`
This adds a variant of the existing `utf8.each_codepoint` function that
additionally passes the index of each codepoint to the given function.
Previously, the only way to iterate through codepoints with indices was
to call `utf8.as_list` first, then iterate through the resulting `List`,
which was a little wasteful.

I wasn't sure whether to call the function `each_with_index` or
`each_codepoint_with_index`. I went with the latter in the end since I
felt it was more consistent/clear, despite being longer.