Allow `List.fold` to take any type for `start`
Previously, `List.fold`'s `start` argument was required to be of the
same type as the inner values of the `List` being operated on. This was
sufficient for some cases, but limited the usefulness of the method. One
example of something that wasn't possible before this change, but is
with it, is calculating a numeric value from a `List` of non-numeric
values:

```
# This calculates the sum of the lengths of each string in a `List` (in
# this case, 11).
["one", "two", "three"].fold(0, (|sum, string| sum + string.size()))
```