`for` loop vars must be created in the proper scope (fixes #620).
Prior to this, `for` temporary vars were always locals. That worked as
long as the temporary wasn't referenced in a lambda. The lambda would
get confused, thinking it needed to become a closure to capture the
local. The lambda's bytecode would attempt to load from a closure that
was never built, crashing the interpreter.

With this commit, temps made by `for` are now in the right scope. The
lambda is thus no longer confused.

It is now possible to reuse global variables within non-ranged `for`.