Introduce vararg transfer via `*value` (fixes #621).
I think this is best explained by example:

```
define sum(a: Integer...): Integer { ... }
var numbers = [1, 2, 3]
sum(*numbers) # [6]
```

Varargs works by collecting excess values into a `List`. When transfer
is used, an existing `List` can instead be used. This was designed to
allow a `Function` to take varargs and pass them down.

Vararg transfer has restrictions. First, this is limited to `List`
values, and to varargs only (this is not splat, and won't be). Transfer
is not limited to vars (`*[value]` is perfectly fine), but it must be
the only excess argument. These restrictions are likely to stay.

For simplicity, this commit omits implementing transfer for calls that
use keyargs.